✨ (network): Almost complete the RakNet part

+++ network/src/main/kotlin/me/hexalite/liteware/network/session/NetworkPlayerSession.kt
@@ -1,12 +1,12 @@
 package me.hexalite.liteware.network.session
 
-import me.hexalite.liteware.network.raknet.LitewareRakNetServer
+import me.hexalite.liteware.network.LitewareRakNetServer
 import me.hexalite.liteware.protocol.ProtocolVersion
 import java.net.SocketAddress
 
 data class NetworkPlayerSession(
     val protocolVersion: ProtocolVersion,
     val address: SocketAddress,
-    val mtu: Int,
+    val mtu: Short,
     val server: LitewareRakNetServer
 )
